初期値と目標値とprogressを用いて現在値を算出する共通メソッドを作成して共通化。座標移動とサイズ変更の両方をprogress更新に追従させた。（まだバグっているが、一歩手前まで行っている）

+++ YoutubeViewer/FloatingImage/FloatingImageWindow.swift
@@ -26,7 +26,7 @@ class FloatingImageWindow: UIWindow {
         removeFromSuperview()
     }
     
-    func updateImageViewFrame(dismissalProgress: Float, tabBarHeight: CGFloat) {
+    func updateImageViewFrame(dismissalProgress: CGFloat, tabBarHeight: CGFloat) {
         if let floatingImageVC = rootViewController as? FloatingImageViewController {
             floatingImageVC.updateImageViewFrame(dismissalProgress: dismissalProgress, tabBarHeight: tabBarHeight)
         }
